Move web-console gem to test-only, to avoid this error mesage:
```
$ bin/rake db:migrate RAILS_ENV=test
Web Console is activated in the test environment. This is
usually a mistake. To ensure it's only activated in development
mode, move it to the development group of your Gemfile:

    gem 'web-console', group: :development

If you still want to run it in the test environment (and know
what you are doing), put this in your Rails application
configuration:

    config.web_console.development_only = false
```